+++ io/web/assets/js/home.js
@@ -1,0 +1,6 @@
+$(document).ready(function(){
+  $("#advanced_options").on("click", function(e) {
+    e.preventDefault();
+    $("#advanced_selections").toggle();
+  });
+});