+++ io/web/templates.go
@@ -1,12 +1,15 @@
 package web
 
 import (
+	"fmt"
 	"html/template"
 	"io"
 	"os"
 	"path"
+	"strings"
 
 	rice "github.com/GeertJohan/go.rice"
+	vlib "github.com/gnames/gnlib/domain/entity/verifier"
 	"github.com/labstack/echo/v4"
 	"github.com/pkg/errors"
 )
@@ -68,10 +71,58 @@ func parseFiles(box *rice.Box, t *template.Template) (*template.Template, error)
 		} else {
 			tmpl = t.New(name)
 		}
+		addFuncs(tmpl)
 		_, err = tmpl.Parse(s)
 		if err != nil {
 			return nil, err
 		}
 	}
 	return t, nil
+}
+
+func addFuncs(tmpl *template.Template) {
+	tmpl.Funcs(template.FuncMap{
+		"classification": func(pathStr, rankStr string) string {
+			if pathStr == "" {
+				return ""
+			}
+			paths := strings.Split(pathStr, "|")
+			var ranks []string
+			if rankStr != "" {
+				ranks = strings.Split(rankStr, "|")
+			}
+
+			res := make([]string, len(paths))
+			for i := range paths {
+				path := strings.TrimSpace(paths[i])
+				if len(ranks) == len(paths) {
+					rank := strings.TrimSpace(ranks[i])
+					if rank != "" {
+						path = fmt.Sprintf("%s (%s)", path, rank)
+					}
+				}
+				res[i] = path
+			}
+			return strings.Join(res, " >> ")
+		},
+		"matchType": func(mt vlib.MatchTypeValue, ed int) template.HTML {
+			var res string
+			clr := map[string]string{
+				"green":  "#080",
+				"yellow": "#a80",
+				"red":    "#800",
+			}
+			switch mt {
+			case vlib.Exact:
+				res = fmt.Sprintf("<span style='color: %s'>%s match</span>", clr["green"], mt)
+			case vlib.NoMatch:
+				res = fmt.Sprintf("<span style='color: %s'>%s</span>", clr["red"], mt)
+			case vlib.Fuzzy, vlib.PartialFuzzy:
+				res = fmt.Sprintf("<span style='color: %s'>%s match, edit distance: %d</span>", clr["yellow"], mt, ed)
+			default:
+				res = fmt.Sprintf("<span style='color: %s'>%s match</span>", clr["yellow"], mt)
+			}
+			return template.HTML(res)
+		},
+	})
 }